Refactored according to review feedback
