--- apps/interpreter/src/visitors/PostgresColumnTypeVisitor.ts
@@ -1,3 +1,0 @@
-import { SQLColumnTypeVisitor } from './SQLColumnTypeVisitor';
-
-export class PostgresColumnTypeVisitor extends SQLColumnTypeVisitor {}
