--- apps/interpreter/src/visitors/PostgresValueRepresentationVisitor.ts
@@ -1,3 +1,0 @@
-import { SQLValueRepresentationVisitor } from './SQLValueRepresentationVisitor';
-
-export class PostgresValueRepresentationVisitor extends SQLValueRepresentationVisitor {}
