--- apps/interpreter/src/visitors/SQLiteColumnTypeVisitor.ts
@@ -1,3 +1,0 @@
-import { SQLColumnTypeVisitor } from './SQLColumnTypeVisitor';
-
-export class SQLiteColumnTypeVisitor extends SQLColumnTypeVisitor {}
