--- apps/interpreter/src/visitors/SQLiteValueRepresentationVisitor.ts
@@ -1,3 +1,0 @@
-import { SQLValueRepresentationVisitor } from './SQLValueRepresentationVisitor';
-
-export class SQLiteValueRepresentationVisitor extends SQLValueRepresentationVisitor {}
