+++ libs/language-server/src/grammar/attributes.langium
@@ -32,6 +32,3 @@ fragment LayoutAttributeFragment:
 
 RuntimeParameter:
   'requires' name=ID;
-
-TableAttribute returns StringAttribute:
-  'table' StringAttributeFragment;
