+++ libs/language-server/src/grammar/blocks/PostgresLoader.langium
@@ -1,4 +1,5 @@
 import '../attributes'
+import 'shared-attributes'
 
 PostgresLoader:
   name='PostgresLoader' '{'
