+++ libs/language-server/src/grammar/blocks/shared-attributes.langium
@@ -1,0 +1,4 @@
+import '../attributes'
+
+TableAttribute returns StringAttribute:
+  'table' StringAttributeFragment;
